+++ spec/spec_helper_acceptance.rb
@@ -1,17 +1,6 @@
-require 'beaker-rspec'
-require 'beaker/puppet_install_helper'
+# frozen_string_literal: true
 
-run_puppet_install_helper
+require 'puppet_litmus'
+PuppetLitmus.configure!
 
-RSpec.configure do |c|
-  proj_root = File.expand_path(File.join(File.dirname(__FILE__), '..'))
-
-  c.formatter = :documentation
-
-  c.before :suite do
-    puppet_module_install(source: proj_root, module_name: 'chrony')
-    hosts.each do |host|
-      on host, puppet('module', 'install', 'puppetlabs-stdlib'), acceptable_exit_codes: [0]
-    end
-  end
-end
+require 'spec_helper_acceptance_local' if File.file?(File.join(File.dirname(__FILE__), 'spec_helper_acceptance_local.rb'))